Fix non-deterministic Pay::Customer#subscription
Postgres doesn't guarantee return order, so this sorts by ID to ensure
the same record is returned each time.

The previous implementation was also trying to optimize the association
and made this more complex than it needed to be.

Users may want to override this to support by `created_at` if they're
using UUIDs instead. This makes it simpler to override and customize.

Closes #647